Move tasks.py
